fix dockerfile so it incrementally builds
The dockerfile was completely out of order. Things that don't often change
from build to build should be placed at the top of the dockerfile and things
that change every build should be placed at the bottom of the dockerfile.
This way, the docker image can incrementally build and it takes less time to
build a new image as opposed to what was happening previously; the image was
being built from scratch each time becasue the copy .. happened at the top.